+++ src/test/java/de/gematik/demis/nps/service/storage/TransactionBundleFactoryTest.java
@@ -45,7 +45,7 @@ import org.mockito.Mockito;
 class TransactionBundleFactoryTest {
 
   private static final String EXPECTED =
-      """
+"""
 {
     "resourceType": "Bundle",
     "type": "transaction",