+++ ReleaseNotes.md
@@ -2,6 +2,14 @@
 
 # Release Notification Processing Service
 
+## Release 2.2.0
+- Introduced Feature Flag to disable the communication with the Pseudonymization Storage Service (PSS)
+- uses lenient parser when relaxed validation was used to send data to NRS
+
+## Release 2.1.1
+- change base chart to istio hostnames
+- updating dependencies
+
 ## Release 2.1.0
 - lift test user configuration up to controller
 - prepare new test user configuration mechanism based on HTTP headers